+++ sqitch/deploy/function_invite_feedback_ids.sql
@@ -4,11 +4,11 @@
 
 BEGIN;
 
-CREATE FUNCTION maevsi_private.invite_feedback_ids() RETURNS TABLE (invitation_feedback_id INTEGER) AS $$
+CREATE FUNCTION maevsi_private.invite_feedback_ids() RETURNS TABLE (invitation_feedback_id BIGINT) AS $$
 BEGIN
     RETURN QUERY
     SELECT invite_account.invitation_feedback_id FROM maevsi.invite_account
-        WHERE   invite_account.account_id = current_setting('jwt.claims.account_id', true)::INTEGER
+        WHERE   invite_account.account_id = current_setting('jwt.claims.account_id', true)::BIGINT
             OR  invite_account.event_id IN (SELECT maevsi.events_organized())
     UNION ALL
     SELECT invite_contact.invitation_feedback_id FROM maevsi.invite_contact
