+++ sqitch/deploy/table_invite_contact.sql
@@ -13,9 +13,9 @@ BEGIN;
 CREATE TABLE maevsi.invite_contact (
     id                        BIGSERIAL PRIMARY KEY,
     uuid                      UUID NOT NULL UNIQUE DEFAULT maevsi.uuid_generate_v1mc(),
-    event_id                  INTEGER REFERENCES maevsi.event(id) NOT NULL,
-    contact_id                INTEGER REFERENCES maevsi.contact(id) NOT NULL,
-    invitation_feedback_id    INTEGER REFERENCES maevsi.invitation_feedback_data(id) NOT NULL,
+    event_id                  BIGINT REFERENCES maevsi.event(id) NOT NULL,
+    contact_id                BIGINT REFERENCES maevsi.contact(id) NOT NULL,
+    invitation_feedback_id    BIGINT REFERENCES maevsi.invitation_feedback_data(id) NOT NULL,
     UNIQUE (event_id, contact_id)
 );
